requirements.txt
fastapi==0.63.0
uvicorn==0.13.3
PyMySQL==0.10.1
sqlalchemy==1.3.22
email-validator==1.1.2
python-dotenv
flake8
autopep8
mysqlclient
starlette-context
fastapi_login
python-multipart
fastapi-cloudauth
boto3
python-jose